Refactor prometheus helpers into separate module
